feat:add database schema
