+++ .idea/dataSources/c7c880b7-8734-4be6-a6ec-f6cc576de146/storage_v2/_src_/database/postgres.edMnLQ.meta
@@ -1,0 +1,1 @@
+#n:postgres
