+++ .idea/dataSources/c7c880b7-8734-4be6-a6ec-f6cc576de146/storage_v2/_src_/database/postgres.edMnLQ/schema/pg_catalog.0S1ZNQ.meta
@@ -1,0 +1,2 @@
+#n:pg_catalog
+!<md> [null, 0, null, null, -2147483648, -2147483648]
