+++ .idea/sqlDataSources.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="DdlMappings">
+    <mapping uuid="cd38aa16-9694-4133-bf78-ae3e5a0e0660" name="libraryManagementSys Mapping">
+      <data-sources db="e0edf9fb-aabf-4891-9ec0-552903beaf47" ddl="0e599311-d0a7-4d3e-9c41-eeef6153b4d3" />
+    </mapping>
+  </component>
+</project>
