+++ src/main/java/com/example/librarymanagementsystem/HelloApplication.java
@@ -1,5 +1,6 @@
 package com.example.librarymanagementsystem;
 
+import com.example.librarymanagementsystem.utils.DatabaseConnection;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
@@ -18,6 +19,9 @@ public class HelloApplication extends Application {
     }
 
     public static void main(String[] args) {
+
         launch();
+
+        DatabaseConnection.getConnection();
     }
 }
