+++ src/main/java/com/example/librarymanagementsystem/enums/Status.java
@@ -1,5 +1,5 @@
 package com.example.librarymanagementsystem.enums;
 
 public enum Status {
-    RESERVED, PENDING, BORROWED, RETURNED, OVERDUE
+    AVAILABLE, PENDING, BORROWED, OVERDUE
 }
